+++ addons/docs/src/typings.d.ts
@@ -1,12 +1,9 @@
 declare module '@mdx-js/react';
-declare module '@storybook/addon-docs/mdx-compiler-plugin';
-declare module '@storybook/addon-docs/blocks';
 declare module 'global';
 declare module '@egoist/vue-to-react';
 declare module 'remark-slug';
 declare module 'remark-external-links';
 declare module 'babel-plugin-react-docgen';
 declare module 'require-from-string';
-declare module 'tmp';
-declare module 'cross-spawn';
 declare module 'styled-components';
+declare module 'acorn-jsx';