+++ examples/vue-kitchen-sink/.storybook/preview.js
@@ -1,4 +1,3 @@
-import { addParameters } from '@storybook/vue';
 import Vue from 'vue';
 import Vuex from 'vuex';
 
@@ -7,8 +6,9 @@ import MyButton from '../src/stories/Button.vue';
 Vue.component('my-button', MyButton);
 Vue.use(Vuex);
 
-addParameters({
+export const parameters = {
+  passArgsFirst: true,
   docs: {
     iframeHeight: '60px',
   },
-});
+};