+++ lib/api/src/index.tsx
@@ -9,12 +9,13 @@ import React, {
   useMemo,
   useRef,
 } from 'react';
+import { mergeWith } from 'lodash';
 
 import {
-  SET_STORIES,
   STORY_CHANGED,
   SHARED_STATE_CHANGED,
   SHARED_STATE_SET,
+  SET_STORIES,
 } from '@storybook/core-events';
 import { RenderData as RouterData } from '@storybook/router';
 import { Listener } from '@storybook/channels';
@@ -93,9 +94,14 @@ export type ManagerProviderProps = RouterData &
     children: ReactNode | ((props: Combo) => ReactNode);
   };
 
+// These types are duplicated in addons.
+export type StoryId = string;
+export type StoryKind = string;
+
 export interface Args {
   [key: string]: any;
 }
+
 export interface ArgType {
   name?: string;
   description?: string;
@@ -106,6 +112,19 @@ export interface ArgType {
 export interface ArgTypes {
   [key: string]: ArgType;
 }
+
+export interface Parameters {
+  [key: string]: any;
+}
+
+// This is duplicated from @storybook/client-api for the reasons mentioned in lib-addons/types.js
+export const combineParameters = (...parameterSets: Parameters[]) =>
+  mergeWith({}, ...parameterSets, (objValue: any, srcValue: any) => {
+    // Treat arrays as scalars:
+    if (Array.isArray(srcValue)) return srcValue;
+
+    return undefined;
+  });
 
 export type ModuleFn = (m: ModuleArgs) => Module;