+++ lib/components/src/controls/types.ts
@@ -1,8 +1,11 @@
+import { ArgType } from '../blocks';
+
 /* eslint-disable @typescript-eslint/no-empty-interface */
 export interface ControlProps<T> {
   name: string;
   value: T;
   defaultValue?: T;
+  argType?: ArgType;
   onChange: (name: string, value: T) => T | void;
 }
 
@@ -46,15 +49,15 @@ export type OptionsControlType =
   | 'select'
   | 'multi-select';
 
-export type OptionsConfig = {
+export interface OptionsConfig {
   options: Options;
   controlType: OptionsControlType;
-};
+}
 
-export type NormalizedOptionsConfig = {
+export interface NormalizedOptionsConfig {
   options: OptionsObject;
   controlType: OptionsControlType;
-};
+}
 
 export type TextValue = string;
 export interface TextConfig {}